BE/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
pydantic
loguru
typing-extensions
Pillow
pdf2image
python-docx
opencv-python-headless
numpy
pandas
openpyxl
torch
transformers
tqdm
scikit-image
paddleocr
paddlepaddle
realesrgan
segment-anything
python-dotenv
